+++ db/sqlc/transfers.sql.go
@@ -20,9 +20,9 @@ INSERT INTO transfers (
 `
 
 type CreateTransfersParams struct {
-	FromAccountID int64   `json:"from_account_id"`
-	ToAccountID   int64   `json:"to_account_id"`
-	Amount        float64 `json:"amount"`
+	FromAccountID int64 `json:"from_account_id"`
+	ToAccountID   int64 `json:"to_account_id"`
+	Amount        int64 `json:"amount"`
 }
 
 func (q *Queries) CreateTransfers(ctx context.Context, arg CreateTransfersParams) (Transfer, error) {
